fix ALterado SaleDTO SalesFinalPrice para double

+++ TakingTest/TakingTest.Application/DTO/SaleDTO.cs
@@ -22,6 +22,6 @@ namespace TakingTest.Application.DTO
         public List<SalesProductDTO> SaleProducts { get; set; }
         [JsonIgnore]
         [IgnoreDataMember]
-        public long SalesFinalPrice { get; set; }
+        public double SalesFinalPrice { get; set; }
     }
 }
